requirements.txt
anthropic>=0.25.0
supabase==2.4.0
google-api-python-client>=2.120.0
gspread>=6.0.0
google-auth>=2.28.0
python-telegram-bot>=21.0
pytrends>=4.9.2
pydantic>=2.6.0
python-dotenv>=1.0.0
aiohttp>=3.9.0